Get room from model instead of from app object
Even if the same ParticipantListView can be reused with different rooms
each of its items are shown only for a participant in a specific room
(as the participant models are created again whenever a new room is set
in the ParticipantCollection). Therefore there is no need to get the
active room from the app object; it can be got directly from its model
when the item view is created.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>