Do not show the CollectionsView when the participant list is shown
The CollectionsView should be shown only for logged in users, but not
for guest users. Moreover, as the CollectionsView file is not included
in the guest JavaScript files an error happens if it is tried to be
shown in that case.

Due to this, instead of showing the CollectionsView in
"_showParticipantList" it is now done in its own "_showCollectionsView",
which is called only for logged in users.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>