Fix flickering spinner on avatars during calls with the MCU
The MCU sends the nick of each participant again and again, which caused
the avatar to be continuously set, and this caused a loading icon to
appear on the avatar. Now the avatar (and the name) is set again only if
the user id or the participant name (including the "Guest" placeholder)
has changed since it was last set.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>