Add Babel presets to make Vue compatible with IE11
Otherwise the main Talk UI, the Talk settings and the projects in the
sharing tab of the Files app do not work in Internet Explorer 11.

The Babel presets were copied from the ones used for the server
settings; from now on the presets will be automatically taken into
account when the Vue code is built again by running "make".

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>