Fix handling of WebRTC supported but "getUserMedia" not available
"OCA.SpreedMe.webrtc.capabilities.support" is true when both
"RTCPeerConnection" and "getUserMedia" are available. However, in some
cases WebRTC may be supported by the browser, yet "getUserMedia" may be
not available if the code is not run in a secure context (like HTTPS).
Due to this, now the "WebRTC is not supported in your browser" message
is shown only if "RTCPeerConnection" is not available; if
"RTCPeerConnection" is available but "getUserMedia" is not then it will
be handled as a "NotAllowedError" thrown by "getUserMedia" when called
in an unsecure context.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>