Fix stopping screensharing from the browser
In the old LocalMedia module bundled in the official SimpleWebRTC 2.2.3
the local screen was not removed before "localScreenStopped" was
emitted. However, in the current LocalMedia module it is, so
"SimpleWebRTC.stopScreenShare" was never called when the screensharing
was stopped from the browser (instead of from the Talk UI itself), and
thus the Talk UI was not updated neither locally nor for the remote
peers.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>