Override screen sharing support from "webrtcsupport" module
The "webrtcsupport" module does not take into account if the standard
"getDisplayMedia" is available, and it only sees Firefox and
Chromium/Chrome as supported; screen sharing in Edge was reported as
supported only because it was mismatched for Chrome. Now the default
checks are overridden to match the behaviour of the custom
"getScreenMedia" module.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>