Do not split the screen sharing menu in Firefox >= 66
Before the standard "getDisplayMedia" was supported in Firefox a custom
API was used for screen sharing, which made possible to choose whether
to show a window or the screen. However, limiting the available display
sources with constraints is explicitly forbidden in the
"getDisplayMedia" spec, so that feature was dropped from Firefox when
the standard API started to be used, and thus the screen sharing menu
should be no longer split.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>