Fix sending signaling messages without payload
This reverts commit c02c8c089e67989383e7a90dd0a423343caeae85.

Not all call messages contain a payload attribute, so those that do not
contain it fail to be sent due to "data.payload" being undefined.
Moreover, the missing JSON attributes only happens due to the type of
the payload object used for offers and answers, so it is better fixed
(in a following commit) where they are passed to the signaling.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>