Fix sending signaling messages with unserialized attributes
When offers and answers are created they are sent in the payload of a
signaling message to the other peer. Both offers and answers are
RTCSessionDescription objects which only serialize its own attributes,
so even if the nick was added to the objects it was never serialized and
sent. Now, instead of sending extended RTCSessionDescription objects a
regular object with a copy of the needed attributes is sent.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>