Fix condition to ignore stats in "checkMediaPeer"
Once a result was found all the following stats that had "bytesReceived"
were checked, even if their media type did not match. However, note
that, while incorrect, this did not cause any practical issue, as the
first time that the peer was unmuted the proper stat was checked, and
the following incorrect checks just unmuted again an already unmuted
peer.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>