Fix vertical position of confirm icon in authenticate page
The server explicitly sets the height of confirm icons inside password
inputs, which overrides the height set in Talk for submit inputs in the
authenticate page. Therefore, the height set by the server needs to be
overriden to honour the height set in Talk (which matches the height of
the password input in the authenticate page).

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>