Do not rebuild JavaScript files before packing a release
As the JavaScript files are versioned again they should not be rebuilt
before packing a release; otherwise the JavaScript files in the package
could be different than the ones in the version control system.

As the JavaScript files are no longer built it is not needed either to
set up the development environment; however, it is still necessary to
remove the build directory before packaging the release, so this is now
explicitly done.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>