Add switch to send or not to send video when it is available
Until now, when local video was available it was always sent to the
other peers (directly, or indirectly through the MCU); even if the video
was disabled a black stream was sent. Despite being more lightweight
than the full video that black stream still needs to be processed by the
clients, which could require lots of resources when several of those
black streams are sent in a large group call.

The proper way to handle this would be not to send the video when it is
disabled and only send it when it is available and enabled, triggering a
renegotiation of the streams whenever the video is enabled or disabled.
However, renegotiation is not supported yet in the WebUI nor in mobile
apps, so for the time being a switch was added that, when disabled,
prevents the video from being sent even if it is available.

Note that this switch is currently "static", that is, Peer objects are
not updated when the switch changes; they must be destroyed and created
again for the value to be taken into account.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>