Add naive forced reconnect with the internal signaling server
Forced reconnections with the internal signaling server just leave the
call and join it again. However, as the same session as before is used,
the peers must be explicitly ended before the reconnection is forced;
otherwise when the list of participants is refreshed the same peers as
before will be in there and no reconnection would happen.

Even if the reconnection to the other peers is caused by them being
removed note that leaving the call and joining it again is needed too;
without this the WebRTC session id (Peer.sid) would not change, so the
reconnecting peer would not be seen as stale by the remote peers and
thus they would just soft reconnect, without recreating the Peer object.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>