Force reconnection if video is enabled when it is not being sent
If the video is available but not being sent a forced reconnection is
needed to start sending it, as there is no support for renegotiation and
the Peer objects need to be created again.

The same call flags as the current ones will be used when joining the
call again due to a forced reconnection, so those flags need to be
updated to specify that video became available.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>