Do not send video by default in rooms with 5 or more than 5 participants
Now when joining a call in a room with 5 or more than 5 participants the
video is disabled and not sent by default, even if available. Enabling
the video will cause a reconnection to the other peers and from that
point on the video will be sent, even if it is later disabled. If the
call is left and then joined again the video will be not sent until
explicitly enabled again (unless the number of participants in the
conversation is now below 5).

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>