Move handling of failures pulling messages outside of signaling object
When pulling messages from the internal signaling server fails several
times to pulling of messages is stopped. This is the only thing that
should be done in the signaling object; other handling like leaving the
current room or leaving the call in WebRTC should be done in their
respective objects as a response to an event triggered by the signaling
object.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>