Make rename button always visible
Rename buttons for guest participants and rooms should always be visible
(at least with opacity) to make their very existence more obvious.

Before they were always hidden and shown with full opacity when the
cursor was on the label wrapper, even if it was possible to interact
with the button only when the cursor was on it. Now, as the buttons are
always visible with a low opacity, the full opacity is set only when the
cursor is on the button itself and thus when it can be interacted with.

Signed-off-by: Sascha Wiswedel <sascha.wiswedel@nextcloud.com>