Fix call flags not kept after joining a pending call
When the external signaling server is used "joinCall()" does not join
the call if its room has not been joined first. However, when that
happens, the call is automatically joined later once the room has been
joined. In that case the same flags given when "joinCall()" was called
should be used once the call is actually joined; otherwise, if no flags
are provided, the default flags would be used by the server, which may
not match the real call flags.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>