Fix joining call before joining room after forced reconnection
When the external signaling server is used, after successfully sending a
"joinRoom" message to the Nextcloud server a message is sent too the
external signaling server (Standalone._joinRoomSuccess); this registers
the Nextcloud session in the signaling server, and is needed by the
signaling server to be able to properly process further messages from
that session.

Joining a call requires that the Nextcloud session is registered, so
"joinCall()" is queued until joining the room if called before that;
whether the session is registered or not is tracked with
"signalingRoomJoined".

When a reconnection is forced the room is joined again and, if the
participant was in a call, the call is joined again. As the Nextcloud
session changed the call should not be joined until the new session is
registered again. However, as "signalingRoomJoined" was emptied only in
"leaveRoom()" and this is not called in a forced reconnection its value
still matched the room token, so the call was immediately joined after a
reconnection without properly waiting for the session to be registered.
Due to this the external signaling server could received the message to
join the call after or before the session was registered (depending on
how fast the messages were sent and processed), and thus the call was
properly joined or not depending on the case.

In the external signaling server the Nextcloud session is unregistered
when the "bye" message is received, so for parallelism with that
behaviour "signalingRoomJoined" is now emptied too when sending the
"bye" message (instead of explicitly emptying it only for forced
reconnections).

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>