Show warning when browser silently fails to get user media
In some cases the browser may enter in a faulty state in which
"getUserMedia" does not return neither successfully nor with an error.
It is not possible to detect this except by guessing when some time
passes and the user has not granted nor rejected the media permissions,
so now a warning is shown to the user when that happens.

However, not having granted nor rejected the media permissions after
some time does not guarantee that there is a problem; it could simply be
that the user has not done it yet. Therefore, the warning is hidden if
the media permissions are finally granted or rejected, or when joining
the call has been interrupted (for example, due to changing to a
different room).

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>