Fix message list not reloaded after being shown again
When the virtual list is hidden adding new elements to it has no effect;
it is necessary to explicitly reload it once visible for the new
elements to be loaded.

During a call the chat is moved to a tab in the sidebar, and when a
different tab is selected the previous one is detached (and, thus,
hidden). Therefore, when the chat tab is shown again the message list
must be reloaded to load the messages that could have been added while
hidden.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>