Fix view for participants without streams
When the MCU is used if a participant has no audio nor video streams no
peer connection will be created by that participant to send the streams
to the MCU, and no peer connections will be created by the rest of the
participants to receive the streams of that participant from the MCU. As
a VideoView is created for every participant when they join the call and
later updated once the peer connection is established the VideoView for
participants without streams was always kept in its initial state. Now
the VideoView is explicitly set to a "connected without streams" state
as soon as it is created for participants without streams.

Note, however, that this shows the avatar and the "audio not available"
icon, but it does not show the participant name (except a generic
"Guest" for guests), as the participant name is updated through the data
channel of the peer connection, so when the participant has no streams
the nick is never updated nor even set.

All the issues described above happen also if no MCU is used and both
the local participant and the remote one have no streams; if the local
participant or the remote one has any stream then a peer connection is
created between them, even if the other participant has no stream.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>