Do not set guest name if joined room is not current room
This is more a formal fix, as for guests a "joinRoom" event would not be
triggered with a room that is not the current room, but if it was
trying to set the name would fail.

Besides that it accomodates further changes in the "joinRoom" handler,
as it will be typically used when the current room is joined.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>