Fix message dates in public share authentication page
The chat messages are expected to have the full width of its content
row, except for some space used for the date. However, as chat messages
are left padded that padding must be included too in the width;
otherwise there could be no room for the date.

This is only a problem in the public share authentication page, as apps
use by default "box-sizing: border-box" in all its elements.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>