Use shared Browserslist config for Babel in root
Browserslist is used by Babel, which in turn is used in Talk to bundle
the SimpleWebRTC code with its dependencies (like WebRTC adapter). Until
recently, in the server its configuration was got from the Babel config,
but now it is stated in the "package.json" file. However, it is not
hardcoded, but provided through a new NPM module,
"browserslist-config-nextcloud". Browserslist gets its configuration
from the "package.json" in current or parent directories, so if
SimpleWebRTC was bundled and the "spreed" directory was a subdirectory
of the "server" directory of Nextcloud Browserslists got its
configuration from the "package.json" file in the server. In that case
the bundle failed because "browserslist-config-nextcloud" was not part
of the modules installed for Talk development.

To prevent that, now "browserslist-config-nextcloud" is a development
dependency of Talk. Also, for consistency,
"browserslist-config-nextcloud" is now always used when bundling
SimpleWebRTC, so the result is the same whether the Talk directory is a
subdirectory of the server or not when performing the bundle.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>