Get target browsers for Babel in "vue/" from browserslist
If the target browsers are not explicitly set in the Babel configuration
they will be got from the browserslist configuration. The browserslist
configuration for the "vue/" directory is set in its "package.json" file
with the same targets that there were in the Babel configuration, so
removing them from the Babel configuration does not change anything at
this point. However, this is a preparatory step to keep the target
browsers in sync with the ones used in Nextcloud server by means of the
browserslist configuration.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>