Prevent chat from being compressed in narrow screens
When the sidebar is shown the app content is compressed to make room for
the sidebar. In narrow screens (and specially in very narrow screens)
this broke the appearance of the app content (unreadable chat, unusable
call view), so now the app content is forced to keep the full viewport
width on narrow screens; this causes the sidebar to slide over the app
content instead of compressing it.

Signed-off-by: Marco Ambrosini <marcoambrosini@pm.me>