Keep last visible message when new message field height changes
The new message field is below the message list, so adding or removing
lines from the message being composed changes the height of the new
message field and, with it, the height of the message list. This change,
however, did not affect the scroll position of the message list, so
adding lines to the new message field caused it to overlap the last
visible message in the list. Now the height change in the new message
field is handled to update the scroll position of the message list and
keep the last visible message at the bottom (both when lines are added
and removed from the new message field).

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>