Fixed inconsistent date tooltips in messages
The author row was covering the date on the right side of the chat view
and preventing the date tooltip to be triggered by hovering on the date.
This fix makes the author row slightly narrower (which was the same
approach already used when the date is shown next to a message) in
order to expose the date, thus allowing the date tooltip to be shown
properly.

Signed-off-by: Marco Ambrosini <marcoambrosini@pm.me>