Update VideoView on ICE disconnected/failed with MCU
When the ICE connection state changed to disconnected or failed the
VideoView was not updated when the MCU was used. However, when the MCU
is not used the VideoView is updated to let the user know that there is
a problem in the connection with the other peer. Thus, now the code was
adjusted to update the VideoView regardless of whether the MCU is used
or not.

The only exception is the "FAILED_NO_RESTART" state, as it applies only
when no MCU is used.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>