Return the same deferred object when the room synchronization is pending
When a room synchronization was requested a deferred object was
returned, but when the room synchronization was requested again while
the first request had not finished yet a different deferred object was
returned (the one from the base "syncRooms" method). For consistency,
and to prepare the code for a following change to automatically retry
failed room synchronizations, now the original deferred object is
returned if the room synchronization is requested again while the first
request has not finished yet.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>