Retry failed room synchronization until success
Now when the room synchronization in "internalSyncRooms" fails it is
automatically retried until it succeeds; further calls to
"internalSyncRooms" (or even "Standalone.syncRooms") after a failed
synchronization will just wait for a retried synchronization to succeed.

This is similar to the error handling when pulling chat messages or
trying to reconnect the socket, although the maximum timeout for room
synchronization is lower to try to recover the UI (which greatly depends
on the rooms data) faster after repeated synchronization failures.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>