Force synchronizing rooms on socket reconnection
When the socket is reconnected the rooms are synchronized again.
However, if the synchronization failed before that the new
synchronization could take some time, as it was waiting for the delayed
retry to start. It is very likely that if a synchronization failed it
will succeed once the socket is reconnected (as the socket would have
been disconnected due to network issues), so instead of waiting for the
delayed retry now the rooms are immediately synchronized again when the
socket is reconnected.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>