Make build system compatible with /compile bot
The /compile bot first installs the NPM packages with "npm init" and
then performs the build with "npm run build", which executes the command
defined in "scripts->build" in "package.json".

Currently "npm init" does not install all the NPM packages needed by
Talk; it is necessary to run it both in the root folder and in the
"/vue" folder. Due to this "make npm-init" needs to be run from the
build script to install all the needed packages before performing the
actual build.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>