Ignore offers when no longer in the call
When the current participant lefts a call WebRTC is not properly
tore down; signaling events are still handled, so if an offer message is
received after leaving a call (which in general should not happen, but
in some strange cases it could) a new Peer object was created. Although
the proper fix would be to disconnect WebRTC and the signaling events
after leaving a call that would require deeper changes, so for now
offers received when not in a call are ignored.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>