Cancel delayed peer creation when the other participant leaves the call
When a participant is not in the call a Peer object should not be
created for her; besides being incorrect, that would also create a
VideoView for that participant, which will basically stay there as a
ghost without never connecting to her. Due to this when a participant
leaves the call the delayed peer creation for that participant, if any,
has to be cancelled; otherwise once the timeout expires the peer would
be created with the issue just described.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>