Fix using "activeRoom" in the shortcut handler before joining a room
Note that the fullscreen shortcut could be used in the starting page.
However, that would hide the navigation bar and no room could then be
joined, so it would not make much sense. Therefore, the shortcuts are
totally disabled in the starting page instead of fixing calling
"get('participantFlags')" on a null room.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>