Emit events when the user speaks or stops speaking while muted
The LocalMedia helper handles the local streams sent to the other
participants, and analyzes those streams to emit events when the user
starts and stops speaking. However, when the local audio is muted those
streams are disabled, so the audio is silent and it can not be detected
if the user is speaking. In order to properly analyze the streams when
the local audio is muted the streams are cloned and those clones, which
are never disabled, are now the ones analyzed.

Thanks to this now the LocalMedia helper also emits the
"speakingWhileMuted" and "stoppedSpeakingWhileMuted" events while
respecting the previous behaviour for the "speaking" and
"stoppedSpeaking" events; muting and unmuting the audio while speaking
also emits again the events as needed.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>