Show browser notification if the browser is hidden
Until now an OC.Notification, which shows an element in the page, was
always used when speaking while muted. Now that is used only if the
browser is visible; if it is not that notification would not be seen by
the user, so in that case now a browser notification is used instead.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>