Add volume indicator for the local audio
The volume is shown as a vertical bar next to the audio icon that grows
and shrinks (with a fixed bottom position) depending on the volume
level. The bar is coloured with a gradient from green to yellow to red
for better clarity of the current volume.

The previous way of indicating speech (a pulse animation in the audio
icon) has not been modified; this is an additional visualization (only
for the local audio) to provide a more noticeable feedback to the user.

Both the existing button and the new bar had to be wrapped in a helper
div to include the volume bar in the old layout.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>