Fix audio not played before video is received in Chromium
Chromium does not play the audio in a video element before the video is
available. Due to this, when the remote peer had audio and video
available if both were disabled (for example, when the remote peer has
just joined the call) and then the audio was enabled the audio from the
remote peer was not heard in Chromium until the video was also enabled.
Instead of using a video element for both the audio and the video now an
audio element was added to play the audio tracks of the remote stream.

The change is applied to all browsers, though, as using an audio element
for the audio tracks makes sense in all of them (even if until now they
worked fine except for Chromium).

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>