Request offer again when receiving peer fails
When the MCU is used and the connection fails in the peer that sent the
offer, that is, the own peer object that sends the local media to the
MCU, a forced reconnection is triggered. However, if the peer that
failed was one used to receive the media from the other participants
there was no reconnection, and therefore the other participant could be
no longer heard or seen.

Now, if the connection with one of those peers fail an offer is
requested again to the MCU periodically until the connection is
established again. The same approach is used too in the initial
connection to those peers, as in some cases the initial offer request
may fail, which left the peers disconnected.

The periodic retry is stopped as soon as an offer is received, but even
if that offer fails to connect eventually an ICE fail will be triggered,
which will start a new reconnection routine.

Finally, note that a forced reconnection when the connection to other
participants fail is not only unneeded, but also it would be
problematic, as a network issue in one participant could trigger a
forced reconnection for all the other participants.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>