Fix setters failing in EditableTextLabel before first render
The setter methods in EditableTextLabel (indirectly) update its HTML
elements using "getUI()". Those setters may be used before the label is
first rendered (for example, to set up some default values in the
initialization of a parent view), so "bindUIElements()" needs to be
called to prevent "getUI()" from failing due to "this._ui" not being
defined yet in those cases (note that no UI elements will be actually
bound, as before the first render they have not been created yet, but
that is not a problem due to the way in which the elements are updated
from the setters).

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>