Fix integration test for joining a file room again
The integration tests check that a user can join a file room again after
leaving it, but the user is removed from the room when leaving it due to
how the tests are written; in "the real world" leaving the room does not
remove the user.

The reason is that in the tests the room token is got by one user but
the room is joined by a different user. Currently the user is added to
the room as a permanent participant when getting the token instead of
when joining the room, so in the tests the user that joins the room is a
self-joined user that is removed once she leaves the room; in "the real
world" the user would have needed to get the token of the file room,
which would add her to the room and thus she would not be removed after
leaving, only after explicitly removing herself.

Due to all this the tests were changed to test joining the file room
again after the user removes herself instead of after she leaves the
room.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>