Add user to file room when joining it instead of when getting its token
This fixes the code so it behaves as stated in the documentation.

In practice currently this should make no difference, as in the Files
app the user joins the room immediately after getting its token;
however, this makes possible to get the token without joining the room,
which will be needed in the future to improve the UX of the Files app.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>